refactor: modularize window and download logic
